feat: Substitute component call with component (lazy)
